requirements.txt
python-dotenv
gitpython
langchain
langchain-community
openai
jinja2
click
